added ability to reproduce added age and lifespan variables.  bugs can die from old age now eat() damage and health gain are based on aggression levels. isPredator is now a variable created on initialization
added a new array list called "babyBugs" and moved the die() call to a for loop after everything has run.  The babies are added to the general pool after processing.  This avoids for loop fails.

added a public variable called time.  may consider combining the clock for the hive with this global variable.